% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsegNeuromorphometrics-package.R
\docType{package}
\name{ggsegNeuromorphometrics-package}
\alias{ggsegNeuromorphometrics}
\alias{ggsegNeuromorphometrics-package}
\title{ggsegNeuromorphometrics: Neuromorphometrics Brain Atlas for the 'ggseg' Ecosystem}
\description{
Neuromorphometrics maximum probability tissue labels, derived from the MICCAI 2012 Grand Challenge and Workshop on Multi-Atlas Labeling and distributed with SPM12, for the 'ggseg' ecosystem. Provides cortical and subcortical 'ggseg_atlas' objects with both 2D polygon geometry and 3D vertex indices, for use with 'ggseg' and 'ggseg3d'. The label data are released under CC BY-NC; MRI scans originate from the OASIS project and the labeled data are provided by Neuromorphometrics, Inc. under academic subscription.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggsegverse/ggsegNeuromorphometrics}
  \item Report bugs at \url{https://github.com/ggsegverse/ggsegNeuromorphometrics/issues}
}

}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})

Authors:
\itemize{
  \item Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})
}

}
\keyword{internal}
