% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{miccai_cortical}
\alias{miccai_cortical}
\title{MICCAI Cortical Atlas}
\usage{
miccai_cortical()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\description{
Cortical regions from the Neuromorphometrics maximum probability tissue
labels distributed with SPM12, derived from the MICCAI 2012 Grand Challenge
and Workshop on Multi-Atlas Labeling. Projected onto the fsaverage5
surface.
}
\details{
The label data are released under CC BY-NC. Credit the MRI scans as
originating from the OASIS project and the labeled data as "provided by
Neuromorphometrics, Inc. under academic subscription".
}
\examples{
miccai_cortical()
}
\references{
Landman BA, Warfield SK (Eds.) (2012). \emph{MICCAI 2012 Workshop on
Multi-Atlas Labeling}.

Marcus DS, et al. (2007). Open Access Series of Imaging Studies (OASIS):
cross-sectional MRI data in young, middle aged, nondemented, and demented
older adults. \emph{Journal of Cognitive Neuroscience}, 19(9):1498-1507.
\doi{10.1162/jocn.2007.19.9.1498}
}
\seealso{
Other ggseg_atlases:
\code{\link[=miccai_subcortical]{miccai_subcortical()}}
}
\concept{cortical_atlases}
\concept{ggseg_atlases}
